Fix some types

+++ include/kulbabu_hardware/kulbabu_hardware_motors.h
@@ -2,7 +2,7 @@
 #define KULBABU_ROS_CONTROL__KULBABU_HARDWARE_MOTOR_H
 
 // C++
-//#include <boost/scoped_ptr.hpp>
+#include <boost/scoped_ptr.hpp>
 
 // ROS
 #include <ros/ros.h>
@@ -24,13 +24,13 @@ public:
   virtual ~KulbabuHardwareMotors() {}
 
   /** \brief Set the buffer for sending to the robot hardware. */
-  virtual void setCommand(uint8_t index, uint8_t cmd_perc);
+  virtual void setCommand(uint8_t index, double cmd_perc);
 
   /** \brief Send buffer to the robot hardware. */
   virtual void doCommand();
 
   /** \brief Read the encoder count from the robot hardware. */
-  virtual int getEncoderVelocity(uint8_t index);
+  virtual double getEncoderVelocity(uint8_t index);
 
 protected:
 
@@ -40,7 +40,7 @@ protected:
   // Startup and shutdown of the internal node inside a roscpp program
   ros::NodeHandle nh_;
 
-  uint8_t encoder_max_;
+  int encoder_max_;
 
 };  // class
